Add atlass_helper module so that it simplify the usage of pyglet.TextureAtlas
