+++ UI/ui_renderer.py
@@ -1,8 +1,4 @@
-import pyglet
-from pyglet.gl import *
 from pyglet.graphics import Batch, OrderedGroup
-from pyglet.image import AbstractImage, TextureRegion
-from pyglet.image.atlas import TextureBin
 from pyglet.sprite import Sprite
 
 from Helpers.hit_test import HitTest
@@ -16,7 +12,6 @@ class UIRenderer:
         self._window_size: (int, int) = (0, 0)
         self._main_batch = Batch()
         self._main_group = OrderedGroup(0)
-        self._texture_atlas = TextureBin()
         self._ui_object_list = []
 
     @classmethod
@@ -34,17 +29,6 @@ class UIRenderer:
     @classmethod
     def get_main_group(cls) -> OrderedGroup:
         return cls._instance._main_group
-
-    @classmethod
-    def add_texture(cls, texture: AbstractImage, border: int = 0) -> TextureRegion:
-        texture_region = cls._instance._texture_atlas.add(texture, border)
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST)
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST)
-        return texture_region
-
-    @classmethod
-    def load_image(cls, image_path: str, folder: str = 'Static/Images/', border=0) -> TextureRegion:
-        return cls.add_texture(pyglet.image.load(folder + image_path), border)
 
     @classmethod
     def draw(cls):
